requirements.txt
python-telegram-bot>=13.7,<20.0
rich>=10.7.0
eth-ape>=0.6.7
ape-infura==0.6.1
ape-solidity==0.6.2

